refactor: move hidden config field wrapper into framework
