+++ graph-gateway/src/config.rs
@@ -1,17 +1,15 @@
-use std::ops::Deref;
-use std::str::FromStr;
-use std::{collections::BTreeMap, fmt, path::PathBuf};
+use std::{collections::BTreeMap, path::PathBuf};
 
-use alloy_primitives::{Address, B256, U256};
+use alloy_primitives::{Address, U256};
 use graph_subscriptions::subscription_tier::{SubscriptionTier, SubscriptionTiers};
 use secp256k1::SecretKey;
 use semver::Version;
 use serde::Deserialize;
-use serde_with::{serde_as, DeserializeAs, DisplayFromStr, FromInto};
+use serde_with::{serde_as, DisplayFromStr, FromInto};
 use thegraph::types::UDecimal18;
 use toolshed::url::Url;
 
-use gateway_framework::config::Chain;
+use gateway_framework::config::{Chain, Hidden, HiddenSecretKey};
 
 use crate::indexers::public_poi::ProofOfIndexingInfo;
 
@@ -170,41 +168,4 @@ pub struct Subscriptions {
 pub struct SubscriptionsDomain {
     pub chain_id: u64,
     pub contract: Address,
-}
-
-#[derive(Deserialize)]
-#[serde(transparent)]
-pub struct Hidden<T>(pub T);
-
-impl<T: fmt::Debug> fmt::Debug for Hidden<T> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "HIDDEN")
-    }
-}
-
-impl<T: FromStr> FromStr for Hidden<T> {
-    type Err = T::Err;
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Ok(Self(s.parse()?))
-    }
-}
-
-impl<T> Deref for Hidden<T> {
-    type Target = T;
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-struct HiddenSecretKey;
-impl<'de> DeserializeAs<'de, Hidden<SecretKey>> for HiddenSecretKey {
-    fn deserialize_as<D>(deserializer: D) -> Result<Hidden<SecretKey>, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        let bytes = B256::deserialize(deserializer)?;
-        SecretKey::from_slice(bytes.as_slice())
-            .map(Hidden)
-            .map_err(serde::de::Error::custom)
-    }
 }
